pyproject.toml
[project]
name = "ai-foundry-agent-service-mcp-server"
version = "0.1.0"
description = "Your project description"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "aiohttp>=3.12.12",
    "azure-ai-projects>=1.0.0b11",
    "azure-identity>=1.23.0",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
]
